Itinerary_Agent/requirements.txt
python-dotenv>=1.0.1
fastapi>=0.115.0
google-genai>=1.65.0
httpx>=0.28.1
langgraph>=1.0.9
requests>=2.32.3
uvicorn>=0.34.0
